requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-swagger-ui==3.36.0
requests==2.25.0
python-dotenv==0.15.0
SQLAlchemy==1.3.20